requirements-NEW.txt
pandas == 1.3.5
streamlit == 1.8.0
numpy == 1.21.6
Pillow == 9.1.0
click==7.1.2